+++ src/ivpm/__init__.py
@@ -20,7 +20,6 @@
 #*
 #****************************************************************************
 import os
-#from .ivpm_subprocess import ivpm_popen
 from .pkg_info.pkg_compile_flags import PkgCompileFlags
 from .pkg_info.pkg_info import PkgInfo
 from .pkg_info.pkg_info_rgy import PkgInfoRgy